4-multiplication/src/domain/use-cases/create-table.use-case.ts
export interface CreateTableUseCase{
    execute: (options: CreateTableOptions) => string;
}

export interface CreateTableOptions{
    base: number;
    limit?:number;
}


export class CreateTable implements CreateTableUseCase{

    constructor(){

    }
    //método para ejecutar el caso de uso
    execute({base, limit = 10}:CreateTableOptions){//datos necesarios para crear la tabla de multiplicar
        let texto = '';
        for(let i = 1; i<=limit;i++){
            texto+= `${base} x ${i} = ${base*i}`
            if(i<limit) texto+= '\n';
        }
        return texto;
    }
}